refactor(view): move help + loading + rollback modal rendering into view_modals.go (move-only)
